replace self user
